project-consensus-backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=requirements.txt --pip-args=None requirements.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
django==5.2.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.9.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
psycopg[binary]==3.2.10
    # via -r requirements.in
psycopg-binary==3.2.10
    # via psycopg
sqlparse==0.5.3
    # via django
